+++ src/Core/Exceptions/ContainerException.php
@@ -2,7 +2,9 @@
 
 namespace Niang\Core\Exceptions;
 
+use Psr\Container\ContainerExceptionInterface;
+
 /** Erreur de résolution du conteneur DI : classe introuvable, dépendance circulaire, paramètre manquant. */
-class ContainerException extends \RuntimeException
+class ContainerException extends \RuntimeException implements ContainerExceptionInterface
 {
 }